Move quaternion-matrix conversion to utils.rotations
